requirements.txt
requests>=2.25.0
websockets>=10.0
prompt_toolkit>=3.0.0
pyyaml>=5.1
click>=8.0.0